requirements.txt
discord.py
pymongo
Pillow
dotenv
